modify: application config,DepositDto maxLimit int -> Long | main Page, depositinfo

+++ src/main/java/com/hana/app/data/DepositDto.java
@@ -23,7 +23,7 @@ public class DepositDto {
     private int joinDeny;
     private String joinMember;
     private String etcNote;
-    private int maxLimit;
+    private Long maxLimit;
     private String dclsStrtDay;
     private String dclsEndDay;
     private String finCoSubmDay;
